Rename directory 'formatter' -> 'formatters'.
We have multiple after all.
