+++ lib/cff/formatters/bibtex_formatter.rb
