+++ lib/cff/formatters/formatter.rb
